+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # unreleased
 
 - Fix issue where events may be emitted before extension is ready to accept them
+- Prevent extension from loading when using Braintree JS SDK v3.13.0 or older (closes #5)
 
 # 0.2.0