+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Changelog
 
-# unreleased
+# 0.3.0
 
 - Fix issue where events may be emitted before extension is ready to accept them
 - Prevent extension from loading when using Braintree JS SDK v3.13.0 or older (closes #5)